requirements.txt
torch
torchvision
torchaudio
transformers
gradio
langchain

pytest
pylint
black
python-decouple

fastapi
uvicorn
python-multipart

Pillow
pytesseract